Use board[] for non-spoiler descriptions of slays.
Replace do_cmd_script with a boring version.

Replace do_compress_string() with a more generic try() macro.

Create a specific TR1_X15_DRAGON flag for the slay on Lightning. This is not
given specifically to anything else, but can be given to randarts if they are
given both the x3 and x5 slays by the program. I don't know if this is good or
bad.

Rearrange init1.c so that the X15_DRAGON in a_info.txt is parsed correctly.