Remove almost all of the places where the index of an ego item is significant and allow ego items to be added, removed and changed with relative impunity.
Add ego_item_type.min_obj and .max_obj to indicate which objects it can be
applied to. Remove .slot and .level, and replace .rarity with .chance to indicate
how likely it is to be applied relative to other similar ego types.
Remove object_type.xtra1, and replace object_type.xtra2 with .activation (which is
only ever used for that purpose). The other things xtra1 was used for are now
added directly to object_type.flags*.
Rearrange k_info.txt a bit so that each ego type can be added to a single range
of objects.
Add a "rating" field to object_kind, and give a boost to rating based on this for
non-artefacts.
Use the "extra" field in k_info.txt to indicate which armour is modified on
creation in an unusual way (dragon shields/helms and elven cloaks at present).
Set object_type.activation for planar weapons, and avoid checking .art_name when
deciding whether to look for activations.
Mention normal objects if and only if they increase the level rating.
Move the check against cheat_peek inside object_mention().
Add a "special" field to ego_item_type, and use that to define the random features
of ego items rather than using the ego indices directly.
Call the appropriate a_m_aux_* function when every object is created, as it removes
some duplication without changing the effect.
Remove the slight bias towards early elements for most of the EGO_XTRA_* things.
Define one_in() as a slightly clearer version of !rand_int().